Report raw value from HX711 to overcome auto-tare functionality.
Currently, when the device is switched off (intentionally or due to
power cut), Tasmota only supports starting with the last known weight
(Sensor34 7) or starting with 0 kg. In either case, this is sometimes
not the desired functionality. For off-site logging of weights, it is
more useful to report the raw, absolute value from the HX711 and to do
calibration/tare in upper layers.

Note: although it is called "Raw" value, the value is both averaged
and scaled! It is just the offset compensation that is missing.