PR Template. Move the CI Pass request to a Note
As the CI Test is done after the PR is submitted, a checkbox for "The code change must pass CI tests" is always left unchecked. It should be better to just add a note that the proposed PR is going to be tested and that the CI Tests must be passed in order to merge the PR. Any CI Test error should be addressed by the author of the PR, before merging.