Fix detection of SHT40
A SHT40 was not detected by the previous scheme of scanning the I2C
addresses and for each address probing for SHT3x and SHT4x immediately
one after the other. The SHT40 is succesfully detected when first
probing all adresses for SHT3x and then scanning all adresses again for
SHT4x. The fix implements this scheme. It was tested with SHTC3 and
SHT40 sensors.